example-it: renamed package to .performance
